Consolidate overlays into "chaco.overlays" (#741)
* copy modules into chaco/overlays

* make original modules stubs raising deprecation warnings

* copy layers into overlays

* fully populate chaco.overlays.api and chaco.api

* flake8

* update tox.ini by running flake8 -q

* somehow this is also needed in tox.ini but I did not see this flake8 error locally

* remove bad trailing comma

* move data label test to sit in chaco/overlays/tests

* add stub modules in chaco/layers importing from new location and adding deprecation warnigns.  Also expose layers in the chaco.overlays.api

* add layers to the chaco.api as well

* old files that were moved have since been changed on master and did not get updated in merge of master

* flake8

* typo

* update deprecation warning messages

* remove stub modules from autogenerated api docs

* remove chaco/layers/data

* import from chaco.plots in overlay modules

* import Legend from chaco.overlays.api
