+++ chaco/layers/api.py
@@ -1,1 +1,20 @@
-from .status_layer import StatusLayer, ErrorLayer, WarningLayer
+# (C) Copyright 2006-2021 Enthought, Inc., Austin, TX
+# All rights reserved.
+#
+# This software is provided without warranty under the terms of the BSD
+# license included in LICENSE.txt and may be redistributed only under
+# the conditions described in the aforementioned license. The license
+# is also available online at http://www.enthought.com/licenses/BSD.txt
+#
+# Thanks for using Enthought open source!
+import warnings
+
+from chaco.overlays.api import ErrorLayer, StatusLayer, WarningLayer
+
+warnings.warn(
+    "Importing from chaco.layers.api is deprecated, please import from "
+    "chaco.api or chaco.overlays.api going forward as chaco/layers will be "
+    "removed in a future release.",
+    DeprecationWarning,
+    stacklevel=2,
+)
