+++ chaco/overlays/tests/test_data_label.py
