+++ ci/edmtool.py
@@ -123,6 +123,15 @@ doc_ignore = {
     "*/tests",
     # The following stub modules were kept for backwards compatibility but
     # will be removed in the next major relaese. Ref: enthought/chaco#748
+    "chaco/colormapped_selection_overlay.py",
+    "chaco/data_label.py",
+    "chaco/lasso_overlay.py",
+    "chaco/layers/*",
+    "chaco/legend.py",
+    "chaco/plot_label.py",
+    "chaco/scatter_inspector_overlay.py",
+    "chaco/text_box_overlay.py",
+    "chaco/tooltip.py",
     "chaco/barplot.py",
     "chaco/candle_plot.py",
     "chaco/cmap_image_plot.py",
