+++ tox.ini
@@ -6,7 +6,6 @@ exclude =
     docs/source/conf.py
     chaco/plots/quiverplot.py
     chaco/abstract_colormap.py
-    chaco/junk2.py
     chaco/plot_containers.py
     chaco/grid.py
     chaco/base_1d_plot.py
@@ -23,22 +22,14 @@ exclude =
     chaco/transform_color_mapper.py
     chaco/default_colormaps.py
     chaco/plotscrollbar.py
-    chaco/simple_plot_frame.py
     chaco/abstract_plot_renderer.py
     chaco/ticks.py
-    chaco/chaco_plot_editor.py
-    chaco/base_plot_frame.py
     chaco/base_candle_plot.py
-    chaco/thingy.py
     chaco/plots/polygon_plot.py
-    chaco/text_box_overlay.py
-    chaco/legend.py
     chaco/linear_mapper.py
-    chaco/serializable.py
     chaco/abstract_overlay.py
     chaco/selectable_legend.py
     chaco/abstract_plot_data.py
-    chaco/colormapped_selection_overlay.py
     chaco/point_data_source.py
     chaco/plots/colormapped_scatterplot.py
     chaco/plots/jitterplot.py
@@ -49,7 +40,6 @@ exclude =
     chaco/plot_factory.py
     chaco/plot_component.py
     chaco/data_label.py
-    chaco/lasso_overlay.py
     chaco/_speedups_fallback.py
     chaco/plots/multi_line_plot.py
     chaco/label.py
@@ -65,7 +55,6 @@ exclude =
     chaco/color_mapper.py
     chaco/plots/line_scatterplot_1d.py
     chaco/polar_mapper.py
-    chaco/scatter_inspector_overlay.py
     chaco/axis_view.py
     chaco/base.py
     chaco/plots/color_bar.py
@@ -80,7 +69,6 @@ exclude =
     chaco/tools/range_selection.py
     chaco/tools/zoom_tool.py
     chaco/tools/lasso_selection.py
-    chaco/tools/drag_tool.py
     chaco/tools/line_inspector.py
     chaco/tools/traits_tool.py
     chaco/tools/image_inspector_tool.py
@@ -96,21 +84,18 @@ exclude =
     chaco/tools/line_segment_tool.py
     chaco/tools/range_selection_2d.py
     chaco/tools/regression_lasso.py
-    chaco/tools/base_zoom_tool.py
     chaco/tools/range_selection_overlay.py
     chaco/tools/legend_highlighter.py
     chaco/tools/pan_tool2.py
     chaco/tools/better_selecting_zoom.py
-    chaco/tools/simple_zoom.py
     chaco/tools/legend_tool.py
-    chaco/tools/toolbars/toolbar_buttons.py
     chaco/tools/toolbars/plot_toolbar.py
-    chaco/layers/api.py
-    chaco/layers/status_layer.py
-    chaco/layers/svg_range_selection_overlay.py
+    chaco/tools/toolbars/toolbar_buttons.py
     chaco/overlays/simple_inspector_overlay.py
     chaco/overlays/databox.py
     chaco/overlays/api.py
+    chaco/overlays/layers/api.py
+    chaco/overlays/layers/svg_range_selection_overlay.py
     chaco/tests/test_plotcontainer.py
     chaco/tests/test_image_plot.py
     chaco/tests/test_data_view.py
@@ -119,7 +104,6 @@ exclude =
     chaco/tests/test_array_or_none.py
     chaco/tests/test_speedups.py
     chaco/tests/test_base_utils.py
-    chaco/tests/test_serializable.py
     chaco/scales/time_scale.py
     chaco/scales/formatters.py
     chaco/scales/scales.py
@@ -133,6 +117,7 @@ exclude =
     examples/demo/quiver.py
     examples/demo/scales_test.py
     examples/demo/qt_example.py
+    examples/demo/chaco_trait_editor.py
     examples/demo/multiaxis_using_Plot.py
     examples/demo/vertical_plot.py
     examples/demo/stacked_axis.py
@@ -154,7 +139,6 @@ exclude =
     examples/demo/advanced/scalar_image_function_inspector.py
     examples/demo/advanced/javascript_hover_tools.py
     examples/demo/advanced/cmap_variable_sized_scatter.py
-    examples/demo/advanced/scalar_image_function_inspector_old.py
     examples/demo/zoomed_plot/wav_to_numeric.py
     examples/demo/zoomed_plot/zoom_overlay.py
     examples/demo/zoomed_plot/zoom_plot.py
